API Reference

Units

UnitfulGaussian.Fr — Constant

UnitfulGaussian.Fr

The franklin, the CGS-ESU and Gaussian unit of charge.

Dimension: 𝐌^1/2 𝐋^3/2 𝐓^-1.

See also: Unitful.C

UnitfulGaussian.statA — Constant

UnitfulGaussian.statA

The statampere, the CGS-ESU and Gaussian unit of current, equal to 1 franklin per second.

Dimension: 𝐌^1/2 𝐋^3/2 𝐓^-2.

See also: Unitful.A

UnitfulGaussian.statV — Constant

UnitfulGaussian.statV

The statvolt, the CGS-ESU and Gaussian unit of electric potential.

Dimensions: 𝐌^1/2 𝐋^1/2 𝐓^-1.

See also: Unitful.V

UnitfulGaussian.G — Constant

UnitfulGaussian.G

The gauss, the CGS-EMU and Gaussian unit of magnetic B field.

Dimensions: 𝐌^1/2 𝐋^3/2 𝐓^-1.

See also: Unitful.T

Warning

Unitful.jl also defines a gauss as a 10⁻⁴ T. This is not dimensionally equivalent to the unit described here.

UnitfulGaussian.Mx — Constant

UnitfulGaussian.Mx

The maxwell, the CGS-EMU and Gaussian unit of magnetic flux (flux of the magnetic B-field).

Dimensions: 𝐌^1/2 𝐋^3/2 𝐓^-1.

See also: Unitful.Wb

Warning

Unitful.jl also defines a maxwell as 10⁻⁸ Wb. This is not dimensionally equivalent to the unit described here.

UnitfulGaussian.Oe — Constant

UnitfulGaussian.Oe

The oersted, the CGS-EMU and Gaussian unit of the magnetic H field. It is dimensionally equivalent to the gauss (G). The corresponding SI unit for the H field is A/m.

Dimensions: 𝐌^1/2 𝐋^3/2 𝐓^-1.

See also: Unitful.A

Warning

Unitful.jl also defines a oersted as 1000/4π A/m. This is not dimensionally equivalent to the unit described here.

UnitfulGaussian.statF — Constant

UnitfulGaussian.statF

The statfarad, the CGS-ESU and Gaussian unit of capacitance. It is equal to 1cm, and the cm should be used in the Gaussian system of units for capacitance.

Dimensions: 𝐋

See also: Unitful.F

UnitfulGaussian.statH — Constant

UnitfulGaussian.statH

The stathenry, the CGS-ESU and Gaussian unit of inductance. It is equal to 1s^2/cm.

Dimensions: 𝐋^-1 𝐓^2

See also: Unitful.H

Equivalences

UnitfulGaussian.ChargeEquivalence — Type

ChargeEquivalence <: Equivalence

Equivalence type for converting between franklins and coulombs (charge), and between statamperes and amperes (current).

UnitfulGaussian.ElectricFieldEquivalence — Type

ElectricFieldEquivalence <: Equivalence

Equivalence type for converting between V/m and statV/cm (E-field), and between C/m² and Fr/cm² (D-field).

UnitfulGaussian.ElectricFluxEquivalence — Type

ElectricFluxEquivalence <: Equivalence

Equivalence type for converting between V⋅m and statV⋅cm (electric field flux), and between coulombs and franklins (displacement field flux).

UnitfulGaussian.PotentialEquivalence — Type

PotentialEquivalence <: Equivalence

Equivalence type for converting between volts and statvolts (potential difference).

UnitfulGaussian.MagneticFieldEquivalence — Type

MagneticFieldEquivalence <: Equivalence

Equivalence type for converting between tesla and gauss (B-field), and between A/m and oersted (H-field).

UnitfulGaussian.MagneticFluxEquivalence — Type

MagneticFluxEquivalence <: Equivalence

Equivalence type for converting between weber and maxwell (B-field flux).